requirements/requirements-base.txt
Django>=1.7.5
python-etcd>=0.4.1
python-dateutil>=2.2
